% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_get.R
\name{fcr_get}
\alias{fcr_get}
\title{Calculate Feed Conversion Ratio (FCR)}
\usage{
fcr_get(adg_res, adfi_res, quiet = FALSE)
}
\arguments{
\item{adg_res}{A list object returned by the \code{adg_get()} function, containing ADG data and related information.}

\item{adfi_res}{A list object returned by the \code{adfi_get()} function, containing ADFI data and related information.}

\item{quiet}{Logical; if TRUE, suppresses all messages and warnings, default is FALSE.}
}
\value{
A list containing:
\itemize{
\item \code{fcr_res}: A data.table with FCR calculation results
\item \code{adg_info}: The input ADG information table
\item \code{adg_data}: The original ADG data
\item \code{adfi_info}: The input ADFI information table
\item \code{adfi_data}: The original ADFI data
\item \code{fcr_summary}: Statistical summary of FCR calculation results
}
}
\description{
Calculates Feed Conversion Ratio (FCR) based on Average Daily Gain (ADG) and Average Daily Feed Intake (ADFI) data.
This function merges ADG and ADFI data, calculates corrected FCR values, and generates statistical summaries.
}
\details{
The \code{fcr_get} function performs the following steps:
\itemize{
\item Validates the format of input ADG and ADFI data
\item Merges ADG and ADFI data, matching by animal ID and location
\item Calculates corrected daily gain and feed conversion ratio
\item Generates statistical summaries of FCR, including sample size, minimum, maximum, median, mean, standard deviation, and coefficient of variation
}
}
\examples{
result_nedap <- preprocess_data(data = mintyr::nedap, station_type = "nedap", quiet = TRUE)
adg_result <- adg_get(result_nedap, my_break = c(30,120), quiet = TRUE)
adfi_result <- adfi_get(result_nedap, adg_result, quiet = TRUE)
fcr_result <- fcr_get(adg_res = adg_result, adfi_res = adfi_result, quiet = TRUE)
head(fcr_result$fcr_res)
head(fcr_result$fcr_summary)
}
\seealso{
\code{\link{adg_get}} for calculating Average Daily Gain data
\code{\link{adfi_get}} for calculating Average Daily Feed Intake data
}
